backend/requirements.txt
# Django大规模用户架构 - 依赖包
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2

# 数据库
psycopg2-binary==2.9.7
django-extensions==3.2.3

# 缓存和会话
django-redis==5.4.0
redis==4.5.4

# 异步任务
celery==5.3.4
celery[redis]==5.3.4

# 认证和安全
djangorestframework-simplejwt==5.3.0
django-otp==1.2.2
cryptography==41.0.7

# 文件存储
django-storages==1.14.2
boto3==1.29.7

# API文档
drf-spectacular==0.26.5

# 性能监控
django-silk==5.0.4
psutil==5.9.6

# 数据处理
pandas==2.1.3
numpy==1.25.2

# 支付集成
requests==2.31.0

# 开发工具
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0